Move type-related classes into org.spine3.type package.
