+++ core-java/src/main/java/org/spine3/internal/EventHandlerMethod.java
@@ -28,10 +28,10 @@ import com.google.common.collect.Multimap;
 import com.google.protobuf.Message;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
-import org.spine3.EventClass;
 import org.spine3.base.EventContext;
 import org.spine3.eventbus.Subscribe;
 import org.spine3.server.MultiHandler;
+import org.spine3.type.EventClass;
 import org.spine3.util.MethodMap;
 import org.spine3.util.Methods;
