+++ core-java/src/main/java/org/spine3/server/error/CommandHandlerAlreadyRegisteredException.java
@@ -19,7 +19,7 @@
  */
 package org.spine3.server.error;
 
-import org.spine3.CommandClass;
+import org.spine3.type.CommandClass;
 
 /**
  * Exception that is thrown when more than one handler
