+++ core-java/src/main/java/org/spine3/server/internal/CommandHandlerMethod.java
@@ -28,12 +28,12 @@ import com.google.common.collect.Multimap;
 import com.google.protobuf.Message;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
-import org.spine3.CommandClass;
 import org.spine3.Internal;
 import org.spine3.base.CommandContext;
 import org.spine3.internal.MessageHandlerMethod;
 import org.spine3.server.Assign;
 import org.spine3.server.MultiHandler;
+import org.spine3.type.CommandClass;
 import org.spine3.util.MethodMap;
 import org.spine3.util.Methods;
